requirements.txt
line-bot-sdk 
bs4 
flask 
pymongo 
datetime 
pandas 
SnowNLP 
emoji 
pyshorteners 
scipy
ollama
google.cloud
dotenv
os
re
sentence_transformers